requirements.txt
#Theorderofpackagesissignificantbecausepipprocessesthemintheorder
#ofappearance.Changingtheorderhasanimpactontheoverallintegration
#processwhichmaycausewedgesinthegatelater.

python-babel>=1.3
python-cryptography>=0.8
python-dogpile.cache>=0.5.3
python-eventlet>=0.16.1
python-greenlet>=0.3.2
python-iso8601>=0.1.9
python-jsonschema>=2.0.0
python-keystoneclient
python-keystonemiddleware>=1.5.0
python-kombu>=2.4.8
python-ldap>=2.4
python-ldappool>=1.0
python-migrate>=0.9.1
python-msgpack>=0.4.0
python-netaddr>=0.7.12
python-oauthlib>=0.6
python-oslo-concurrency>=1.8.0
python-oslo-config
python-oslo-db>=1.7.0
python-oslo-i18n>=1.5.0
python-oslo-log>=1.0.0
python-oslo-messaging>=1.8.0
python-oslo-middleware>=1.0.0
python-oslo-policy>=0.3.1
python-oslo-serialization>=1.4.0
python-oslo-utils>=1.4.0
python-oslo-context>=0.2.0
python-pam>=0.1.4
python-passlib
python-paste
python-pastedeploy>=1.5.0
python-pbr>=0.6
python-posix-ipc
python-pycadf>=0.8.0
python-pysaml2
python-routes>=1.12.3
python-six>=1.9.0
python-sqlalchemy>=0.9.7
python-webob>=1.2.3
